@@ -1,14 +1,16 @@
 {
   "name": "mqtt-parent-control",
-  "version": "2.0.3",
+  "version": "3.0.0",
   "description": "",
   "main": "index.js",
+  "type": "module",
   "scripts": {
     "build": "tsc",
-    "start": "node ./build/index.js",
+    "start": "node ./src/index.ts",
     "restart": "npm run build && npm start",
     "clean": "rm tmp/* -f",
     "lint": "eslint . --ext .ts",
+    "lint:fix": "eslint . --fix",
     "test": "jest",
     "test:update": "jest --updateSnapshot"
   },
@@ -16,26 +18,28 @@
   "license": "MIT",
   "dependencies": {
     "async-mqtt": "2.6.3",
-    "config": "3.3.8",
-    "csv-parse": "5.3.1",
-    "dayjs": "1.11.6",
-    "debug": "4.3.4",
-    "mqtt": "4.3.7",
-    "p-timeout": "4.1.0"
+    "config": "4.1.1",
+    "csv-parse": "6.1.0",
+    "dayjs": "1.11.18",
+    "debug": "4.4.3",
+    "mqtt": "4.3.8",
+    "p-timeout": "6.1.4"
   },
   "devDependencies": {
-    "@types/config": "3.3.0",
-    "@types/jest": "29.2.2",
-    "@types/node": "17.0.23",
-    "@types/ws": "8.5.3",
-    "@typescript-eslint/eslint-plugin": "5.17.0",
-    "@typescript-eslint/parser": "5.17.0",
-    "eslint": "8.12.0",
+    "@eslint/eslintrc": "3.3.1",
+    "@eslint/js": "9.35.0",
+    "@types/jest": "30.0.0",
+    "@types/node": "24.4.0",
+    "@types/ws": "8.18.1",
+    "@typescript-eslint/eslint-plugin": "8.43.0",
+    "@typescript-eslint/parser": "8.43.0",
+    "eslint": "9.35.0",
     "eslint-config-airbnb-base": "15.0.0",
-    "eslint-config-airbnb-typescript": "16.2.0",
-    "eslint-plugin-import": "2.25.4",
-    "jest": "29.3.0",
-    "ts-jest": "29.0.3",
-    "typescript": "4.8.4"
+    "eslint-config-airbnb-typescript": "18.0.0",
+    "eslint-plugin-import": "2.32.0",
+    "eslint-plugin-stylistic": "0.0.0-0",
+    "jest": "30.1.3",
+    "ts-jest": "29.4.1",
+    "typescript": "5.9.2"
   }
 }
