@@ -3,7 +3,7 @@ import type { OnConnectCallback, OnErrorCallback, OnMessageCallback } from 'mqtt
 
 export interface IMQTTAdapter {
   connected: boolean;
-  publish(topic: string, message: string | Buffer): Promise<void>;
+  publish(topic: string, message: string): Promise<void>;
   subscribe(topic: string | string[] | ISubscriptionMap): Promise<ISubscriptionGrant[]>;
   on (event: 'connect', cb: OnConnectCallback): void
   on (event: 'message', cb: OnMessageCallback): void
