@@ -1,15 +1,15 @@
 import timers from 'timers/promises';
-import * as mqtt from 'async-mqtt';
+import mqtt from 'async-mqtt';
 // @ts-ignore
 import configModule from 'config';
 import pTimeout from 'p-timeout';
 
-import tasks from './tasks';
-import { DummyMqttClient } from './DummyMqttClient';
+import tasks from './tasks/index.ts';
+import { DummyMqttClient } from './DummyMqttClient.ts';
 
-import type { IMQTTAdapter } from './IMQTTAdapter';
-import type Task from './tasks/Task';
-import type { TasksConfig, TaskType } from './tasks/Task';
+import type { IMQTTAdapter } from './IMQTTAdapter.ts';
+import type Task from './tasks/Task.ts';
+import type { TasksConfig, TaskType } from './tasks/Task.ts';
 
 type Config = {
   mqtt: Array<{
@@ -28,7 +28,7 @@ const config = configModule as unknown as Config;
 
 async function withFallBack(task: Task, fn: Function, logs: Array<string>) {
   try {
-    await pTimeout(fn.apply(task), 30_000);
+    await pTimeout(fn.apply(task), {milliseconds: 30_000});
     // @ts-ignore
   } catch (err: Error) {
     logs.push(`${task.name}: ${err.message} ${err.stack}`);
@@ -39,8 +39,8 @@ async function getClient(logs: Array<string>): Promise<IMQTTAdapter> {
   for (let n = 0; n < 3; n++) {
     for (let i = 0; i < config.mqtt.length; i++) {
       try {
-        const client = await pTimeout(mqtt.connectAsync(config.mqtt[i].url, config.mqtt[i].options), 10_000);
-        return client;
+        const client = await pTimeout(mqtt.connectAsync(config.mqtt[i].url, config.mqtt[i].options), {milliseconds: 10_000});
+        return client as IMQTTAdapter;
       } catch (err) {
         logs.push(`MQTT connect failed on attempt ${n} to ${config.mqtt[i].url}`, (err as Error).toString());
       }
@@ -51,6 +51,10 @@ async function getClient(logs: Array<string>): Promise<IMQTTAdapter> {
 }
 
 async function run() {
+  if(!config?.mqtt?.length) {
+    console.log('No mqtt config found');
+    return;
+  }
   let logs:Array<string> = [];
   const client: IMQTTAdapter = await getClient(logs);
   const { tasksConfig } = config;
@@ -84,6 +88,6 @@ async function run() {
   await timers.setTimeout(2000); // wait for may-be-shutdown
 }
 
-pTimeout(run(), 55_000)
+pTimeout(run(), {milliseconds: 55_000})
   .then(() => process.exit(0))
   .catch((err) => { console.log(err); process.exit(1); });
